requirements.txt
absl-py==2.1.0
alembic==1.13.2
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.25.0
idna==3.10
# ipykernel==6.29.5
# ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.5.0
keras-layer-normalization==0.16.0
keras-tuner==1.4.7
kiwisolver==1.4.7
kt-legacy==1.0.5
libclang==18.1.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
# networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
optuna==4.0.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
pure_eval==0.2.3
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
# pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.8.1
safetensors==0.4.5
scikit-datasets==0.2.4
scikit-learn==1.5.2
# scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
# tensorflow-intel==2.17.0
# tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torch-utils==0.1.2
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
